Updated the enemy class and made some minor fixes
First of all the first enemy was still spawning at the edges as it the logic for it in the constructor was not updated, so updated that logic. Also made dependencies less by calling all height and width parameters in the construction of the enemy instance rather than in the logic function seperately. Finally added calling of the enemy character in the drawGameWindow function